msg window support reasoning_content
